+++ t/101-xml.t
@@ -12,6 +12,7 @@ use SemiXML;
 #
 my $filename = 't/test-file.sxml';
 spurt( $filename, q:to/EOSX/);
+#
 $|html [
   $|body [                                       # Body
     $|h1 [ Data from file \# h1 ]                # outside h1
@@ -19,6 +20,7 @@ $|html [
       $|tr [ trrr                                # trrr
         $|th[header \# th ]                      # outside th
         $|td[data \# td ]                        # outside td
+        $|td [! # inside protected body !]
       ]
     ]
   ]
@@ -34,8 +36,11 @@ my Str $xml-text = ~$x;
 ok $xml-text ~~ m/\<html\>/, 'Top level html found';
 ok $xml-text !~~ m/\<head\>/, 'Head not found';
 ok $xml-text ~~ ms/Data from file/, 'Section text found';
-todo 'comments are not implemented yet', 1;
+#todo 'comments are not implemented yet', 1;
 unlike $xml-text, /:s '#' 'outside' 'h1' /, 'comment removed';
+unlike $xml-text, /:s '#' 'trrr' /, 'comment also removed';
+like $xml-text, /:s '#' 'inside' 'protected' 'body' /, 'comment not removed';
+like $xml-text, /:s 'header' '#' 'th' /, 'escaped # not removed';
 
 #say $xml-text;